6.5 - 6.8
