+++ anecdotes_redux/src/components/Anecdotes.js
@@ -1,0 +1,26 @@
+import React from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { vote } from "../reducers/anecdoteReducer";
+
+const Anecdotes = () => {
+  const anecdotes = useSelector((state) => state);
+  const dispatch = useDispatch();
+
+  anecdotes.sort((a, b) => b.votes - a.votes);
+  return (
+    <div>
+      <h2>Anecdotes</h2>
+      {anecdotes.map((anecdote) => (
+        <div key={anecdote.id}>
+          <div>{anecdote.content}</div>
+          <div>
+            has {anecdote.votes}
+            <button onClick={() => dispatch(vote(anecdote.id))}>vote</button>
+          </div>
+        </div>
+      ))}
+    </div>
+  );
+};
+
+export default Anecdotes;
